@@ -19,4 +19,4 @@ flutter:
   uses-material-design: true
 
 environment:
-  sdk: '>=2.16.0 <3.0.0'
+  sdk: '>=2.18.0 <4.0.0'
