@@ -27,22 +27,22 @@ flutter:
 dependencies:
   device_info_plus_platform_interface: ^7.0.0
   ffi: ^2.0.1
-  file: ^6.0.0
+  file: ^6.1.4
   flutter:
     sdk: flutter
   flutter_web_plugins:
     sdk: flutter
-  meta: ^1.3.0
-  win32: ">=2.7.0 <5.0.0"
+  meta: ^1.8.0
+  win32: ^4.1.3
   win32_registry: ^1.1.0
 
 dev_dependencies:
   flutter_lints: ^2.0.1
   flutter_test:
     sdk: flutter
-  mockito: ^5.0.0
-  test: ^1.16.4
+  mockito: ^5.4.0
+  test: ^1.22.0
 
 environment:
-  sdk: ">=2.16.0 <3.0.0"
-  flutter: ">=2.11.0"
+  sdk: ">=2.18.0 <4.0.0"
+  flutter: ">=3.3.0"
