@@ -4,7 +4,7 @@ version: 1.2.3+4
 publish_to: 'none'
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.18.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -13,7 +13,7 @@ dependencies:
   package_info_plus: ^3.1.2
 
 dev_dependencies:
-  build_runner: ^2.0.3
+  build_runner: ^2.3.3
   integration_test:
     sdk: flutter
   flutter_driver:
@@ -21,7 +21,7 @@ dev_dependencies:
   flutter_lints: ^2.0.1
   flutter_test:
     sdk: flutter
-  mockito: ^5.0.7
+  mockito: ^5.4.0
 
 flutter:
   uses-material-design: true
