@@ -29,9 +29,9 @@ dependencies:
     sdk: flutter
   flutter_web_plugins:
     sdk: flutter
-  http: ^0.13.0
-  meta: ^1.3.0
-  path: ^1.8.0
+  http: ^0.13.5
+  meta: ^1.8.0
+  path: ^1.8.2
   package_info_plus_platform_interface: ^2.0.1
   win32: ^4.0.0
 
@@ -39,8 +39,8 @@ dev_dependencies:
   flutter_lints: ^2.0.1
   flutter_test:
     sdk: flutter
-  test: ^1.21.1
+  test: ^1.22.0
 
 environment:
-  sdk: ">=2.16.0 <3.0.0"
-  flutter: ">=2.11.0"
+  sdk: ">=2.18.0 <4.0.0"
+  flutter: ">=3.3.0"
